Added player turns
made SwitchSides a static function that Tile accesses after successful or unsuccessful promotions.
SwitchSides is also called if a promotion is not to be prompted at all.

Enemy Piece promotions behave as expected (though this should be removed in the future due to player not controlling enemy movement/promotions/drops)